fix: long was castes to int when set in statement support for nan and inf at least in postgresql driver including unittest
git-svn-id: svn://svn.code.sf.net/p/tntdb/code/trunk@229 8aebba32-5d12-0410-a889-bbfed4e15866

+++ tntdb/include/tntdb/bits/statement.h
@@ -276,11 +276,11 @@ namespace tntdb
         { stmt.setBool(name, data); }
       void setInt(int data)
         { stmt.setInt(name, data); }
-      void setLong(int data)
+      void setLong(long data)
         { stmt.setLong(name, data); }
       void setUnsigned(unsigned data)
         { stmt.setUnsigned(name, data); }
-      void setUnsignedLong(unsigned data)
+      void setUnsignedLong(unsigned long data)
         { stmt.setUnsignedLong(name, data); }
       void setInt32(int32_t data)
         { stmt.setInt32(name, data); }
